requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
qdrant-client>=2.8.0
httpx>=0.25.0
python-dotenv>=1.0.0
watchdog>=3.0.0
tree-sitter>=0.20.0
tree-sitter-javascript>=0.20.0
gitpython>=3.1.0
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
